type cast value to integer (task #2857)

+++ src/FieldHandlers/IntegerFieldHandler.php
@@ -15,7 +15,7 @@ class IntegerFieldHandler extends BaseFieldHandler
      */
     public function renderValue($table, $field, $data, array $options = [])
     {
-        $result = filter_var($data, FILTER_SANITIZE_NUMBER_INT);
+        $result = (int)filter_var($data, FILTER_SANITIZE_NUMBER_INT);
 
         if (!empty($result) && is_numeric($result)) {
             $result = number_format($result);
